+++ oxfmt.config.ts
@@ -1,0 +1,4 @@
+import { defineConfig } from 'oxfmt'
+import config from 'oxlint-config-alexandernanberg/oxfmt/base'
+
+export default defineConfig(config)